Update README.md
After the docker is up, we may need to log in to the app for first time. Those credentials were hidden inside the file and keeping it in README.md may be useful for new contributors like me. Also the docker-compose yaml, comes bundled with frappe bench therefore the following guidelines on installing the frappe bench may be irrelevant.